Replaced socketService with socket-server, place to root
