--- template/apps/api/src/services/cloud-storage/cloud-storage.helper.ts
@@ -1,8 +1,0 @@
-export const getFileKey = (url: string | null | undefined) => {
-  if (!url) return '';
-
-  const decodedUrl = decodeURI(url);
-  const { pathname } = new URL(decodedUrl);
-
-  return pathname.substring(1);
-};
