+++ template/apps/api/src/services/cloud-storage/cloud-storage.service.ts
@@ -13,8 +13,6 @@ import { Upload } from '@aws-sdk/lib-storage';
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
 import { readFile } from 'node:fs/promises';
 
-import * as helpers from './cloud-storage.helper';
-
 import config from '@/config';
 import { BackendFile, ToCamelCase } from '@/types';
 import { caseUtil } from '@/utils';
@@ -32,6 +30,15 @@ const client = new S3Client({
 const bucket = config.CLOUD_STORAGE_BUCKET;
 
 type UploadOutput = ToCamelCase<CompleteMultipartUploadCommandOutput>;
+
+const getFileKey = (url: string | null | undefined) => {
+  if (!url) return '';
+
+  const decodedUrl = decodeURI(url);
+  const { pathname } = new URL(decodedUrl);
+
+  return pathname.substring(1);
+};
 
 const upload = async (fileName: string, file: BackendFile): Promise<UploadOutput> => {
   const params: PutObjectCommandInput = {
@@ -109,7 +116,7 @@ const deleteObject = async (fileName: string): Promise<DeleteOutput> => {
 };
 
 export default {
-  ...helpers,
+  getFileKey,
   upload,
   uploadPublic,
   getObject,
