+++ template/apps/api/src/services/index.ts
@@ -2,6 +2,5 @@ import analyticsService from './analytics/analytics.service';
 import cloudStorageService from './cloud-storage/cloud-storage.service';
 import emailService from './email/email.service';
 import * as googleService from './google/google.service';
-import socketService from './socket/socket.service';
 
-export { analyticsService, cloudStorageService, emailService, googleService, socketService };
+export { analyticsService, cloudStorageService, emailService, googleService };
