+++ template/apps/api/src/socket-server.ts
@@ -4,11 +4,36 @@ import { Server } from 'socket.io';
 
 import { COOKIES } from 'app-constants';
 
-import socketHelper from './socket.helper';
-
 import logger from '@/logger';
 import pubClient, { redisErrorHandler } from '@/redis-client';
 import validateAccessToken from '@/resources/tokens/methods/validate-access-token';
+
+const getCookie = (cookieString: string, name: string) => {
+  const value = `; ${cookieString}`;
+  const parts = value.split(`; ${name}=`);
+  if (parts && parts.length === 2) {
+    const part = parts.pop();
+    if (!part) {
+      return null;
+    }
+
+    return part.split(';').shift();
+  }
+
+  return null;
+};
+
+const checkAccessToRoom = (roomId: string, data: { userId: string }) => {
+  const [roomType, ...rest] = roomId.split('-');
+  const id = rest.join('-');
+
+  switch (roomType) {
+    case 'user':
+      return id === data.userId;
+    default:
+      return false;
+  }
+};
 
 export default (server: http.Server) => {
   const io = new Server(server);
@@ -24,7 +49,7 @@ export default (server: http.Server) => {
   io.use(async (socket, next) => {
     if (!socket.handshake.headers.cookie) return next(new Error('Cookie not found'));
 
-    const accessToken = socketHelper.getCookie(socket.handshake.headers.cookie, COOKIES.ACCESS_TOKEN);
+    const accessToken = getCookie(socket.handshake.headers.cookie, COOKIES.ACCESS_TOKEN);
 
     const token = await validateAccessToken(accessToken);
 
@@ -42,7 +67,7 @@ export default (server: http.Server) => {
   io.on('connection', (socket) => {
     socket.on('subscribe', (roomId: string) => {
       const { userId } = socket.data;
-      const hasAccessToRoom = socketHelper.checkAccessToRoom(roomId, { userId });
+      const hasAccessToRoom = checkAccessToRoom(roomId, { userId });
 
       if (hasAccessToRoom) {
         socket.join(roomId);
